Add files via upload
For this project, I first started by using a dataset found on Kaggle.com, which included property sale prices across multiple states and regions. The dataset also lacked good categorical variables, and for these reasons, I decided to switch to a more relevant dataset that just looked at recent home sales prices in the Ames, IA area.
With this new dataset, I put together and will attempt to predict the sale price of a home in Ames. I was able to use https://beacon.schneidercorp.com to query for all recent home sales in the Ames area. I queried results for the last 12 months, ensuring we were working with only the most recent data and included a little over 500
different sale records. The three main questions that I hope to be able to answer using this dataset are:
1. Which feature influences the price the most?
2. What is the predicted sale price for an active listing of a home in the Ames area?
3. Which style of home has the highest average sale price?